+++ src-docs/src/views/pagination/indeterminate.tsx
@@ -4,15 +4,13 @@ import { EuiPagination } from '../../../../src/components';
 
 export default function () {
   const [activePage, setActivePage] = useState(0);
-  const PAGE_COUNT = 0;
 
   return (
     <EuiPagination
       aria-label="Indeterminate example"
-      pageCount={PAGE_COUNT}
+      pageCount={0}
       activePage={activePage}
       onPageClick={(activePage) => setActivePage(activePage)}
-      compressed
     />
   );
 }